+++ src/main/kotlin/icu/windea/pls/config/config/CwtConfigService.kt
@@ -75,12 +75,12 @@ object CwtConfigService {
      * @see CwtInjectedConfigProvider.injectConfigs
      */
     @Optimized
-    fun injectConfigs(parentConfig: CwtMemberConfig<*>, configs: MutableList<CwtMemberConfig<*>>): Boolean {
+    fun injectConfigs(parentConfig: CwtMemberConfig<*>, containerConfig: CwtMemberConfig<*>, configs: MutableList<CwtMemberConfig<*>>): Boolean {
         var r = false
         val eps = CwtInjectedConfigProvider.EP_NAME.extensionList
         eps.forEachFast f@{ ep ->
             if (!ep.supports(parentConfig)) return@f
-            r = r || ep.injectConfigs(parentConfig, configs)
+            r = r || ep.injectConfigs(parentConfig, containerConfig, configs)
         }
         return r
     }