+++ src/main/kotlin/icu/windea/pls/ep/config/config/CwtConfigPostProcessor.kt
@@ -2,12 +2,15 @@ package icu.windea.pls.ep.config.config
 
 import com.intellij.openapi.extensions.ExtensionPointName
 import icu.windea.pls.config.config.CwtMemberConfig
+import icu.windea.pls.model.ParadoxGameType
 
 /**
  * 用于在成员规则解析完毕后，执行额外的处理逻辑。
  */
 interface CwtConfigPostProcessor {
-    fun supports(config: CwtMemberConfig<*>): Boolean = true
+    fun supports(gameType: ParadoxGameType): Boolean = true
+
+    fun supports(config: CwtMemberConfig<*>): Boolean = supports(config.configGroup.gameType)
 
     fun deferred(config: CwtMemberConfig<*>): Boolean = false